+++ Stage1/planning/movement.py
@@ -1,5 +1,5 @@
-from perception.sensors import *
-from environment.enrvironment2d import *
+from perception.cameras import *
+from environment.env_1d import *
 import time
